FIX 15 - JS seperated and README File is there as well
