+++ Index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
     <title>Nadun Avishka Portfolio</title>
     <link rel="stylesheet" href="style.css" />
-    <script defer src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
 </head>
 <body>
     <aside class="sidebar expanded">
