+++ data/projects.json
@@ -27,7 +27,7 @@
 				"source": "drive",
 				"link": "https://drive.google.com/file/d/YourFileID/preview",
 				"format": "horizontal",
-				"stills": []
+				"images": []
 			}
 		}
 	],
@@ -49,7 +49,7 @@
 				"source": "youtube",
 				"link": "https://www.youtube.com/watch?v=K7nV9zwF0nk",
 				"format": "horizontal",
-				"stills": [ "assets/thumbs/Toppi Welenda Thumb.jpg", "assets/thumbs/Toppi Welenda Thumb.jpg", "assets/thumbs/Toppi Welenda Thumb.jpg" ]
+				"images": [ "assets/thumbs/Toppi Welenda Thumb.jpg", "assets/thumbs/Toppi Welenda Thumb.jpg", "assets/thumbs/Toppi Welenda Thumb.jpg" ]
 			}
 		},
 		{
@@ -69,7 +69,7 @@
 				"source": "drive",
 				"link": "https://drive.google.com/file/d/1FRiMEof6Voneic9bmgcwKa38eYZo5FUO/preview",
 				"format": "vertical",
-				"stills": []
+				"images": []
 			}
 		},
 		{
@@ -86,7 +86,7 @@
 				"source": "drive",
 				"link": "https://drive.google.com/file/d/YourFileID/preview",
 				"format": "horizontal",
-				"stills": []
+				"images": []
 			}
 		}
 	]
