+++ style.css
@@ -1,8 +1,8 @@
 :root {
-    --clr-primary: #1CB7A9;
+    --clr-primary: #ffd300;
     --clr-bg: rgba(40,52,61,0.5);
-    --clr-dark: #28343D;
-    --clr-accent: #374457;
+    --clr-dark: #000000;
+    --clr-accent: #8d7817;
     --clr-light: #fff;
     --sidebar-width: 250px;
 }
@@ -359,7 +359,7 @@ body {
     left: 0;
     right: 0;
     height: 56px;
-    background: rgba(40,52,61,0.85);
+    background: rgb(0, 0, 0, 0.85);
     backdrop-filter: blur(12px);
     z-index: 110; /* above sidebar and overlays */
     align-items: center;
